resolving ex001

+++ exercises/ex001/converteCelsiusParaFahrenheit.php
@@ -1,6 +1,6 @@
 <?php
 
-function converteCelsiusParaFahrenheit(float $celsius): float
+function converteCelsiusParaFahrenheit(int $celsius): float
 {
     return (($celsius * 9/5) + 32);
 }
